requirements.txt
# ============================================
# V1 DEPENDENCIES (Original AlignCV)
# ============================================

# Backend Dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# CORS for frontend integration
fastapi-cors==0.0.6

# Frontend (Streamlit)
streamlit==1.28.1

# Semantic matching (Phase 2)
sentence-transformers>=2.7.0
scikit-learn==1.3.2
numpy==1.26.0
torch>=2.1.0

# Document parsing (Phase 2)
PyPDF2==3.0.1
python-docx==1.1.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
requests==2.31.0

# ============================================
# V2 DEPENDENCIES (AlignCV Pro)
# ============================================

# Database & ORM
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Advanced Document Parsing
PyMuPDF==1.23.8
python-docx==1.1.0
spacy==3.7.2

# File Storage
boto3==1.34.10
firebase-admin==6.3.0

# Task Queue & Caching
redis==5.0.1
celery==5.3.4

# Email
sendgrid==6.11.0

# AI & LLM
mistralai==0.1.0
openai==1.6.0

# Vector Database
qdrant-client==1.7.0
sentence-transformers==2.7.0

# Job Scraping
beautifulsoup4==4.12.2
feedparser==6.0.11

# Environment & Config
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Enhanced NLP
transformers==4.36.0

# Additional Testing
httpx==0.25.2
faker==21.0.0
